+++ src/Components/Switch.test.js
@@ -4,7 +4,7 @@ import { render, screen } from '@testing-library/react';
 import Switch from './Switch';
 
 test('Switch Component Renders', () => {
-    render(<Switch data-testid="webapps-Switch-component" />);
+    render(<Switch label="webapps-Switch-component" />);
 
-    expect(screen.getByTestId('webapps-Switch-component')).toBeDefined();
+    expect(screen.getByText('webapps-Switch-component')).toBeDefined();
 });